timesheet.py
This is a script I wrote to put my timesheet at work. It login is to website with username and password, save my timesheet for each day and on Friday it submit my timesheet. Also on friday I will get text message telling with how many hours I have submitted this week.