Create carwash.py
This is a very small script, where it still needs more modification . Basic goal of this script is to let you know when to wash your car based on the forecast weather. Script looks at current weather and next 3 days weather condition. If it's not rain in 3 next days you will get a text message saying "It's a perfect day to wash your car"